@@ -17,84 +22,78 @@
  },
  "license": "UNLICENSED",
  "author": {
-  "name": "Aran Leite de Gusmão",
+  "name": "Aran Leite",
   "email": "hyoretsu@gmail.com",
-  "url": "https://www.linkedin.com/in/aranleite/"
+  "url": "https://www.linkedin.com/in/aranleite"
  },
  "scripts": {
-  "build": "babel src --extensions \".js,.ts\" --out-dir dist --copy-files --no-copy-ignored",
-  "dev:server": "ts-node-dev -r tsconfig-paths/register --inspect --no-notify --respawn --transpile-only --ignore-watch /node_modules/ src/shared/infra/http/server.ts",
-  "preinstall": "npx only-allow pnpm",
-  "start": "ts-node src/shared/infra/http/server.ts",
+  "build": "babel src --extensions .ts --out-dir dist --copy-files --no-copy-ignored --source-maps",
+  "dev": "tsnd -r tsconfig-paths/register --inspect --no-notify --respawn -T -I /node_modules/ src/shared/infra/http/server.ts",
+  "start": "npm run build && node dist/shared/infra/http/server.js",
   "test": "jest --passWithNoTests",
-  "typeorm": "tsnd -r tsconfig-paths/register ./node_modules/typeorm/cli.js"
- },
- "husky": {
-  "hooks": {
-   "pre-commit": "lint-staged"
-  }
+  "typeorm": "tsnd -r tsconfig-paths/register ./node_modules/typeorm/cli.js",
+  "prepare": "husky install"
  },
  "lint-staged": {
   "src/**/*.ts": [
    "eslint --fix",
-   "yarn test --bail --findRelatedTests",
-   "git add"
+   "yarn test --bail --findRelatedTests"
   ]
  },
  "dependencies": {
-  "@babel/cli": "^7.12.10",
-  "@babel/core": "^7.12.10",
-  "@babel/node": "^7.12.10",
-  "@babel/plugin-proposal-class-properties": "^7.12.1",
-  "@babel/plugin-proposal-decorators": "^7.12.12",
-  "@babel/preset-env": "^7.12.11",
-  "@babel/preset-typescript": "^7.12.7",
+  "@babel/cli": "^7.14.5",
+  "@babel/core": "^7.14.6",
+  "@babel/node": "^7.14.7",
+  "@babel/plugin-proposal-class-properties": "^7.14.5",
+  "@babel/plugin-proposal-decorators": "^7.14.5",
+  "@babel/preset-env": "^7.14.7",
+  "@babel/preset-typescript": "^7.14.5",
   "babel-plugin-module-resolver": "^4.1.0",
-  "babel-plugin-transform-typescript-metadata": "^0.3.1",
+  "babel-plugin-transform-typescript-metadata": "^0.3.2",
   "bcryptjs": "^2.4.3",
-  "celebrate": "^13.0.4",
-  "class-transformer": "^0.3.1",
+  "celebrate": "^15.0.0",
+  "class-transformer": "^0.4.0",
   "cors": "^2.8.5",
-  "date-fns": "^2.16.1",
-  "dotenv": "^8.2.0",
+  "date-fns": "^2.22.1",
+  "dotenv": "^10.0.0",
   "express": "^4.17.1",
   "express-async-errors": "^3.1.1",
   "imagemin-overwrite": "^7.0.1",
-  "imagemin-pngquant": "^9.0.1",
+  "imagemin-pngquant": "^9.0.2",
   "multer": "^1.4.2",
-  "pg": "^8.5.1",
+  "pg": "^8.6.0",
   "reflect-metadata": "^0.1.13",
-  "sharp": "^0.27.0",
-  "tsyringe": "^4.4.0",
-  "typeorm": "^0.2.29",
+  "sharp": "^0.28.3",
+  "tsyringe": "^4.6.0",
+  "typeorm": "^0.2.34",
   "uuid": "^8.3.2"
  },
  "devDependencies": {
   "@types/bcryptjs": "^2.4.2",
-  "@types/cors": "^2.8.9",
-  "@types/express": "^4.17.9",
-  "@types/hapi__joi": "^17.1.6",
-  "@types/jest": "^26.0.19",
-  "@types/multer": "^1.4.5",
-  "@types/sharp": "^0.26.1",
-  "@types/uuid": "^8.3.0",
-  "@typescript-eslint/eslint-plugin": "^4.11.1",
-  "@typescript-eslint/parser": "^4.11.1",
-  "eslint": "^7.17.0",
+  "@types/cors": "^2.8.12",
+  "@types/express": "^4.17.13",
+  "@types/hapi__joi": "^17.1.7",
+  "@types/jest": "^26.0.24",
+  "@types/multer": "^1.4.7",
+  "@types/sharp": "^0.28.4",
+  "@types/uuid": "^8.3.1",
+  "@typescript-eslint/eslint-plugin": "^4.28.3",
+  "@typescript-eslint/parser": "^4.28.3",
+  "eslint": "^7.31.0",
   "eslint-config-airbnb-base": "^14.2.1",
-  "eslint-config-prettier": "^7.1.0",
-  "eslint-import-resolver-typescript": "^2.3.0",
-  "eslint-plugin-import": "^2.22.1",
+  "eslint-config-prettier": "^8.3.0",
+  "eslint-import-resolver-typescript": "^2.4.0",
+  "eslint-plugin-import": "^2.23.4",
   "eslint-plugin-import-helpers": "^1.1.0",
-  "eslint-plugin-prettier": "^3.3.0",
-  "husky": "^4.3.6",
-  "jest": "^26.6.3",
-  "lint-staged": "^10.5.3",
-  "prettier": "^2.2.1",
-  "ts-jest": "^26.4.4",
-  "ts-node-dev": "^1.1.1",
-  "tsconfig-paths": "^3.9.0",
-  "typescript": "^4.1.3"
+  "eslint-plugin-prettier": "^3.4.0",
+  "husky": "^7.0.1",
+  "jest": "^27.0.6",
+  "lint-staged": "^11.0.1",
+  "prettier": "^2.3.2",
+  "ts-jest": "^27.0.3",
+  "ts-node-dev": "^1.1.8",
+  "tsconfig-paths": "^3.10.1",
+  "typescript": "^4.3.5"
  },
  "engines": {
   "node": ">=12.18.0"
