@@ -1,7 +1,10 @@
-const { pathsToModuleNameMapper } = require('ts-jest/utils');
-const { compilerOptions } = require('./tsconfig.json');
+/* eslint-disable import/no-extraneous-dependencies */
+import type { Config } from '@jest/types';
+import { pathsToModuleNameMapper } from 'ts-jest/utils';
 
-module.exports = {
+import { compilerOptions } from './tsconfig.json';
+
+const config: Config.InitialOptions = {
  clearMocks: true,
  collectCoverage: true,
  collectCoverageFrom: ['<rootDir>/src/modules/**/services/*.ts'],
@@ -12,3 +15,5 @@ module.exports = {
  testEnvironment: 'node',
  testMatch: ['**/*.spec.ts'],
 };
+
+export default config;
