@@ -1,6 +1,6 @@
 import { container } from 'tsyringe';
 
-import uploadConfig from '@config/upload.ts';
+import uploadConfig from '@config/upload';
 
 import DiskStorageProvider from './implementations/DiskStorageProvider';
 import IStorageProvider from './models/IStorageProvider';
